+++ backend/src/Plugin/Factory.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Neucore\Plugin;
+
+class Factory implements FactoryInterface
+{
+
+}