+++ backend/tests/Functional/Controller/PluginController/plugin3/src/TestService3.php
@@ -6,18 +6,22 @@ declare(strict_types=1);
 
 namespace Tests\Functional\Controller\PluginController;
 
-use Neucore\Plugin\CoreCharacter;
+use Neucore\Plugin\CoreAccount;
 use Neucore\Plugin\Exception;
-use Neucore\Plugin\GeneralPluginInterface;
+use Neucore\Plugin\FactoryInterface;
+use Neucore\Plugin\GeneralInterface;
 use Neucore\Plugin\PluginConfiguration;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Log\LoggerInterface;
 
-class TestService3 implements GeneralPluginInterface
+class TestService3 implements GeneralInterface
 {
-    public function __construct(LoggerInterface $logger, PluginConfiguration $pluginConfiguration)
-    {
+    public function __construct(
+        LoggerInterface $logger,
+        PluginConfiguration $pluginConfiguration,
+        FactoryInterface $factory,
+    ) {
     }
 
     public function onConfigurationChange(): void
@@ -28,11 +32,7 @@ class TestService3 implements GeneralPluginInterface
         string $name,
         ServerRequestInterface $request,
         ResponseInterface $response,
-        CoreCharacter $main,
-        array $characters,
-        array $memberGroups,
-        array $managerGroups,
-        array $roles,
+        ?CoreAccount $coreAccount,
     ): ResponseInterface {
         throw new Exception('Exception from general plugin.');
     }