+++ backend/tests/Functional/Controller/User/PlayerController/plugin/src/TestService.php
@@ -6,8 +6,10 @@ declare(strict_types=1);
 
 namespace Tests\Functional\Controller\User\PlayerController;
 
+use Neucore\Plugin\CoreAccount;
 use Neucore\Plugin\CoreCharacter;
 use Neucore\Plugin\Exception;
+use Neucore\Plugin\FactoryInterface;
 use Neucore\Plugin\ServiceAccountData;
 use Neucore\Plugin\PluginConfiguration;
 use Neucore\Plugin\ServiceInterface;
@@ -17,8 +19,24 @@ use Psr\Log\LoggerInterface;
 
 class TestService implements ServiceInterface
 {
-    public function __construct(LoggerInterface $logger, PluginConfiguration $pluginConfiguration)
+    public function __construct(
+        LoggerInterface $logger,
+        PluginConfiguration $pluginConfiguration,
+        FactoryInterface $factory,
+    ) {
+    }
+
+    public function onConfigurationChange(): void
     {
+    }
+
+    public function request(
+        string $name,
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        ?CoreAccount $coreAccount,
+    ): ResponseInterface {
+        throw new Exception();
     }
 
     public function getAccounts(array $characters): array
@@ -65,23 +83,6 @@ class TestService implements ServiceInterface
     public function getAllAccounts(): array
     {
         throw new Exception();
-    }
-
-    public function request(
-        string $name,
-        ServerRequestInterface $request,
-        ResponseInterface $response,
-        CoreCharacter $main,
-        array $characters,
-        array $memberGroups,
-        array $managerGroups,
-        array $roles,
-    ): ResponseInterface {
-        throw new Exception();
-    }
-
-    public function onConfigurationChange(): void
-    {
     }
 
     public function search(string $query): array