+++ backend/tests/Unit/Service/PluginService/plugin-acc2/src/TestService2.php
@@ -7,8 +7,10 @@ declare(strict_types=1);
 
 namespace Tests\Unit\Service\PluginService\plugin\src;
 
+use Neucore\Plugin\CoreAccount;
 use Neucore\Plugin\CoreCharacter;
 use Neucore\Plugin\Exception;
+use Neucore\Plugin\FactoryInterface;
 use Neucore\Plugin\ServiceAccountData;
 use Neucore\Plugin\PluginConfiguration;
 use Neucore\Plugin\ServiceInterface;
@@ -18,8 +20,24 @@ use Psr\Log\LoggerInterface;
 
 class TestService2 implements ServiceInterface
 {
-    public function __construct(LoggerInterface $logger, PluginConfiguration $pluginConfiguration)
+    public function __construct(
+        LoggerInterface $logger,
+        PluginConfiguration $pluginConfiguration,
+        FactoryInterface $factory,
+    ) {
+    }
+
+    public function onConfigurationChange(): void
     {
+    }
+
+    public function request(
+        string $name,
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        ?CoreAccount $coreAccount,
+    ): ResponseInterface {
+        throw new Exception();
     }
 
     public function getAccounts(array $characters): array
@@ -62,23 +80,6 @@ class TestService2 implements ServiceInterface
     public function getAllPlayerAccounts(): array
     {
         throw new Exception();
-    }
-
-    public function request(
-        string $name,
-        ServerRequestInterface $request,
-        ResponseInterface $response,
-        CoreCharacter $main,
-        array $characters,
-        array $memberGroups,
-        array $managerGroups,
-        array $roles,
-    ): ResponseInterface {
-        throw new Exception();
-    }
-
-    public function onConfigurationChange(): void
-    {
     }
 
     public function search(string $query): array