+++ backend/tests/Unit/Service/PluginService/plugin-both/src/TestBoth.php
@@ -6,9 +6,11 @@ declare(strict_types=1);
 
 namespace Tests\Unit\Service\PluginService\plugin\src;
 
+use Neucore\Plugin\CoreAccount;
 use Neucore\Plugin\CoreCharacter;
 use Neucore\Plugin\Exception;
-use Neucore\Plugin\GeneralPluginInterface;
+use Neucore\Plugin\FactoryInterface;
+use Neucore\Plugin\GeneralInterface;
 use Neucore\Plugin\ServiceAccountData;
 use Neucore\Plugin\PluginConfiguration;
 use Neucore\Plugin\ServiceInterface;
@@ -16,10 +18,13 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Log\LoggerInterface;
 
-class TestBoth implements GeneralPluginInterface, ServiceInterface
+class TestBoth implements GeneralInterface, ServiceInterface
 {
-    public function __construct(LoggerInterface $logger, PluginConfiguration $pluginConfiguration)
-    {
+    public function __construct(
+        LoggerInterface $logger,
+        PluginConfiguration $pluginConfiguration,
+        FactoryInterface $factory,
+    ) {
     }
 
     public function onConfigurationChange(): void
@@ -30,11 +35,7 @@ class TestBoth implements GeneralPluginInterface, ServiceInterface
         string $name,
         ServerRequestInterface $request,
         ResponseInterface $response,
-        CoreCharacter $main,
-        array $characters,
-        array $memberGroups,
-        array $managerGroups,
-        array $roles,
+        ?CoreAccount $coreAccount,
     ): ResponseInterface {
         throw new Exception();
     }