+++ backend/tests/Unit/Service/PluginService/plugin-name/src/TestService.php
@@ -5,8 +5,10 @@ declare(strict_types=1);
 
 namespace Tests\Unit\Service\PluginService\plugin\src;
 
+use Neucore\Plugin\CoreAccount;
 use Neucore\Plugin\CoreCharacter;
 use Neucore\Plugin\Exception;
+use Neucore\Plugin\FactoryInterface;
 use Neucore\Plugin\ServiceAccountData;
 use Neucore\Plugin\PluginConfiguration;
 use Neucore\Plugin\ServiceInterface;
@@ -18,9 +20,25 @@ class TestService implements ServiceInterface
 {
     private PluginConfiguration $pluginConfiguration;
 
-    public function __construct(LoggerInterface $logger, PluginConfiguration $pluginConfiguration)
-    {
+    public function __construct(
+        LoggerInterface $logger,
+        PluginConfiguration $pluginConfiguration,
+        FactoryInterface $factory,
+    ) {
         $this->pluginConfiguration = $pluginConfiguration;
+    }
+
+    public function onConfigurationChange(): void
+    {
+    }
+
+    public function request(
+        string $name,
+        ServerRequestInterface $request,
+        ResponseInterface $response,
+        ?CoreAccount $coreAccount,
+    ): ResponseInterface {
+        throw new Exception();
     }
 
     public function getPluginConfiguration(): PluginConfiguration
@@ -68,23 +86,6 @@ class TestService implements ServiceInterface
     public function getAllPlayerAccounts(): array
     {
         throw new Exception();
-    }
-
-    public function request(
-        string $name,
-        ServerRequestInterface $request,
-        ResponseInterface $response,
-        CoreCharacter $main,
-        array $characters,
-        array $memberGroups,
-        array $managerGroups,
-        array $roles,
-    ): ResponseInterface {
-        throw new Exception();
-    }
-
-    public function onConfigurationChange(): void
-    {
     }
 
     public function search(string $query): array