+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 dd mm yyyy
 
 - **BC Break**: Updated [Neucore plugin](https://github.com/tkhamez/neucore-plugin) to 0.10.0.
+- **BC Break**: `Neucore\Plugin\ServiceInterface::request` is now also called if no user is logged in 
+  (URL `/plugin/{id}/{name}`).
 - Added search for service accounts to *Member Data -> Characters*.
 - Duplicate character search results have been removed.